added solutions to vision transformer
