+++ pytorch/11_attention-mechanisms-and-transformers/7_transformer_architecture/1_solution.py
@@ -1,0 +1,243 @@
+import math
+import pandas as pd
+import torch
+from torch import nn
+from d2l import torch as d2l
+import matplotlib.pyplot as plt
+import time
+
+class PositionWiseFFN(nn.Module):  #@save
+    """The positionwise feed-forward network."""
+    def __init__(self, ffn_num_hiddens, ffn_num_outputs):
+        super().__init__()
+        self.dense1 = nn.LazyLinear(ffn_num_hiddens)
+        self.relu = nn.ReLU()
+        self.dense2 = nn.LazyLinear(ffn_num_outputs)
+
+    def forward(self, X):
+        return self.dense2(self.relu(self.dense1(X)))
+
+class AddNorm(nn.Module):  #@save
+    """The residual connection followed by layer normalization."""
+    def __init__(self, norm_shape, dropout):
+        super().__init__()
+        self.dropout = nn.Dropout(dropout)
+        self.ln = nn.LayerNorm(norm_shape)
+
+    def forward(self, X, Y):
+        return self.ln(self.dropout(Y) + X)
+
+class TransformerEncoderBlock(nn.Module):  #@save
+    """The Transformer encoder block."""
+    def __init__(self, num_hiddens, ffn_num_hiddens, num_heads, dropout,
+                 use_bias=False):
+        super().__init__()
+        self.attention = d2l.MultiHeadAttention(num_hiddens, num_heads,
+                                                dropout, use_bias)
+        self.addnorm1 = AddNorm(num_hiddens, dropout)
+        self.ffn = PositionWiseFFN(ffn_num_hiddens, num_hiddens)
+        self.addnorm2 = AddNorm(num_hiddens, dropout)
+
+    def forward(self, X, valid_lens):
+        Y = self.addnorm1(X, self.attention(X, X, X, valid_lens))
+        return self.addnorm2(Y, self.ffn(Y))
+
+class TransformerEncoder(d2l.Encoder):  #@save
+    """The Transformer encoder."""
+    def __init__(self, vocab_size, num_hiddens, ffn_num_hiddens,
+                 num_heads, num_blks, dropout, use_bias=False):
+        super().__init__()
+        self.num_hiddens = num_hiddens
+        self.embedding = nn.Embedding(vocab_size, num_hiddens)
+        self.pos_encoding = d2l.PositionalEncoding(num_hiddens, dropout)
+        self.blks = nn.Sequential()
+        for i in range(num_blks):
+            self.blks.add_module("block"+str(i), TransformerEncoderBlock(
+                num_hiddens, ffn_num_hiddens, num_heads, dropout, use_bias))
+
+    def forward(self, X, valid_lens):
+        # Since positional encoding values are between -1 and 1, the embedding
+        # values are multiplied by the square root of the embedding dimension
+        # to rescale before they are summed up
+        X = self.pos_encoding(self.embedding(X) * math.sqrt(self.num_hiddens))
+        self.attention_weights = [None] * len(self.blks)
+        for i, blk in enumerate(self.blks):
+            X = blk(X, valid_lens)
+            self.attention_weights[
+                i] = blk.attention.attention.attention_weights
+        return X
+
+class TransformerDecoderBlock(nn.Module):
+    # The i-th block in the Transformer decoder
+    def __init__(self, num_hiddens, ffn_num_hiddens, num_heads, dropout, i):
+        super().__init__()
+        self.i = i
+        self.attention1 = d2l.MultiHeadAttention(num_hiddens, num_heads,
+                                                 dropout)
+        self.addnorm1 = AddNorm(num_hiddens, dropout)
+        self.attention2 = d2l.MultiHeadAttention(num_hiddens, num_heads,
+                                                 dropout)
+        self.addnorm2 = AddNorm(num_hiddens, dropout)
+        self.ffn = PositionWiseFFN(ffn_num_hiddens, num_hiddens)
+        self.addnorm3 = AddNorm(num_hiddens, dropout)
+
+    def forward(self, X, state):
+        enc_outputs, enc_valid_lens = state[0], state[1]
+        # During training, all the tokens of any output sequence are processed
+        # at the same time, so state[2][self.i] is None as initialized. When
+        # decoding any output sequence token by token during prediction,
+        # state[2][self.i] contains representations of the decoded output at
+        # the i-th block up to the current time step
+        if state[2][self.i] is None:
+            key_values = X
+        else:
+            key_values = torch.cat((state[2][self.i], X), dim=1)
+        state[2][self.i] = key_values
+        if self.training:
+            batch_size, num_steps, _ = X.shape
+            # Shape of dec_valid_lens: (batch_size, num_steps), where every
+            # row is [1, 2, ..., num_steps]
+            dec_valid_lens = torch.arange(
+                1, num_steps + 1, device=X.device).repeat(batch_size, 1)
+        else:
+            dec_valid_lens = None
+        # Self-attention
+        X2 = self.attention1(X, key_values, key_values, dec_valid_lens)
+        Y = self.addnorm1(X, X2)
+        # Encoder-decoder attention. Shape of enc_outputs:
+        # (batch_size, num_steps, num_hiddens)
+        Y2 = self.attention2(Y, enc_outputs, enc_outputs, enc_valid_lens)
+        Z = self.addnorm2(Y, Y2)
+        return self.addnorm3(Z, self.ffn(Z)), state
+
+class TransformerDecoder(d2l.AttentionDecoder):
+    def __init__(self, vocab_size, num_hiddens, ffn_num_hiddens, num_heads,
+                 num_blks, dropout):
+        super().__init__()
+        self.num_hiddens = num_hiddens
+        self.num_blks = num_blks
+        self.embedding = nn.Embedding(vocab_size, num_hiddens)
+        self.pos_encoding = d2l.PositionalEncoding(num_hiddens, dropout)
+        self.blks = nn.Sequential()
+        for i in range(num_blks):
+            self.blks.add_module("block"+str(i), TransformerDecoderBlock(
+                num_hiddens, ffn_num_hiddens, num_heads, dropout, i))
+        self.dense = nn.LazyLinear(vocab_size)
+
+    def init_state(self, enc_outputs, enc_valid_lens):
+        return [enc_outputs, enc_valid_lens, [None] * self.num_blks]
+
+    def forward(self, X, state):
+        X = self.pos_encoding(self.embedding(X) * math.sqrt(self.num_hiddens))
+        self._attention_weights = [[None] * len(self.blks) for _ in range (2)]
+        for i, blk in enumerate(self.blks):
+            X, state = blk(X, state)
+            # Decoder self-attention weights
+            self._attention_weights[0][i] = blk.attention1.attention.attention_weights
+            # Encoder-decoder attention weights
+            self._attention_weights[1][i] = blk.attention2.attention.attention_weights
+        return self.dense(X), state
+
+    @property
+    def attention_weights(self):
+        return self._attention_weights
+
+data = d2l.MTFraEng(batch_size=128)
+
+# By a deeper transformer architechture, I am assuming they refer to the number of blocks in 
+# the encoder and decoder. I will try increasing num_blks from 2 to its powers, and keep track 
+# of time taken and final validation_loss.
+
+time_taken_dict = {}
+final_val_loss_dict = {}
+for num_blks in [2,4,8,16]:
+
+    # record start time
+    start_time = time.perf_counter()
+
+    # train with appropriate num_blks
+    num_hiddens, dropout = 256, 0.2
+    ffn_num_hiddens, num_heads = 64, 4
+    encoder = TransformerEncoder(len(data.src_vocab), num_hiddens, ffn_num_hiddens, num_heads, num_blks, dropout)
+    decoder = TransformerDecoder(len(data.tgt_vocab), num_hiddens, ffn_num_hiddens, num_heads, num_blks, dropout)
+    model = d2l.Seq2Seq(encoder, decoder, tgt_pad=data.tgt_vocab['<pad>'],lr=0.001)
+    trainer = d2l.Trainer(max_epochs=30, gradient_clip_val=1, num_gpus=1)
+    trainer.fit(model, data)
+
+    # record end time
+    end_time = time.perf_counter()
+
+    # get the time 
+    time_taken = end_time - start_time
+
+    # get final_val_loss for this num_blk value
+    losses = []
+    for batch in trainer.val_dataloader:
+            with torch.no_grad():
+                Y_hat = model(*batch[:-1])
+                losses.append(model.loss(Y_hat, batch[-1]))
+    final_val_loss = sum(losses)/len(losses)
+
+    # add to dicts
+    final_val_loss_dict[num_blks] = final_val_loss
+    time_taken_dict[num_blks] = time_taken
+
+# now, plot the values
+# Prepare data: sort by key
+block_sizes = sorted(time_taken_dict.keys())
+time_vals = [time_taken_dict[k] for k in block_sizes]
+loss_vals = [final_val_loss_dict[k] for k in block_sizes]
+
+# Create two subplots
+fig, ax1 = plt.subplots(figsize=(7, 4))
+
+# Plot time taken
+ax1.plot(block_sizes, time_vals, 'o-', color='tab:blue', label='Time Taken (s)')
+ax1.set_xlabel('Block Size')
+ax1.set_ylabel('Time Taken (s)', color='tab:blue')
+ax1.tick_params(axis='y', labelcolor='tab:blue')
+ax1.grid(True, which='both', linestyle='--', alpha=0.5)
+
+# Plot loss on a second y-axis
+ax2 = ax1.twinx()
+ax2.plot(block_sizes, loss_vals, 's--', color='tab:red', label='Final Val Loss')
+ax2.set_ylabel('Final Val Loss', color='tab:red')
+ax2.tick_params(axis='y', labelcolor='tab:red')
+
+plt.title('Time Taken and Final Val Loss vs Block Size')
+fig.tight_layout()
+plt.show()
+
+# Conclusion:
+# So time taken seems to increase linearly with num_blocks. And interestingly, validation loss 
+# seems to be increasing as well with more blocks. My guess is that this is due to overfitting
+# in the model with more blocks, and our training data is also less in size and complexity, 
+# which makes this more likely to happen.
+
+
+
+
+### other stuff (ignore)
+temp = next(iter(trainer.val_dataloader))
+type(temp)
+temp[0].shape
+temp[1].shape
+temp[2].shape
+temp[3].shape
+temp2 = [1,2,3]
+sum(temp2)
+data.arrays[0].shape
+data.arrays[0][:5, :5]
+data.arrays[1][:5, :5]
+data.arrays[2].shape
+
+trainer.__dir__()
+model.__dir__()
+model.accuracy()
+type(model.plot())
+trainer.optim.__dir__()
+type(trainer.optim.state)
+len(trainer.optim.state)
+trainer.optim.state.__dir__()
+model.accuracy()
+
